requirements.txt
streamlit>=1.28.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.26.0
pytest>=7.0.0
reportlab>=4.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
celery>=5.3.4
redis>=5.0.1
minio>=7.2.15
requests>=2.31.0